third_party/WebKit/LayoutTests/fast/dom/custom/svg-use-shadow-tree-expected.txt

CONSOLE ERROR: line 2752: Uncaught Error: assert_false: Should not call createdCallback in UA ShadowRoot. expected false got true
CONSOLE ERROR: line 2752: Uncaught Error: assert_false: Should not call attachedCallback in UA ShadowRoot. expected false got true
CONSOLE ERROR: line 2752: Uncaught Error: assert_false: Should not call createdCallback in UA ShadowRoot. expected false got true
CONSOLE ERROR: line 2752: Uncaught Error: assert_false: Should not call attachedCallback in UA ShadowRoot. expected false got true
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_false: Should not call attachedCallback in UA ShadowRoot. expected false got true
PASS SVG <use> shadow trees should not be exposed through custom elements.
Harness: the test ran to completion.
